requirements.txt
streamlit==1.31.0
anthropic==0.18.1
requests==2.31.0
pandas==2.2.0
plotly==5.18.0
python-dotenv==1.0.0
reportlab==4.0.9
Pillow==10.2.0